refactor(providers): rename handle_status_openai_compat and extract http_status module
The status-mapping helpers (`handle_status`, `handle_response`,
`map_http_error_to_provider_error`) are format-agnostic — they parse
both OpenAI and Anthropic error shapes and are used by 12 providers.

- Rename `handle_status_openai_compat` → `handle_status` and
  `handle_response_openai_compat` → `handle_response`.
- Move these helpers from `openai_compatible` into a dedicated
  `providers::http_status` module. `openai_compatible` re-exports
  them for convenience.

Pure refactor, no logic changes.
